requirements.txt
aniso8601==8.0.0
apturl==0.5.2
asgiref==3.2.7
beautifulsoup4==4.4.1
bert-extractive-summarizer==0.4.2
blinker==1.3
blis==0.4.1
boto3==1.12.46
botocore==1.15.46
Brlapi==0.6.4
catalogue==1.0.0
chardet==2.3.0
checkbox-support==0.22
click==7.1.1
command-not-found==0.3
cryptography==1.2.3
cymem==2.0.3
defer==1.0.6
Django==3.0.5
django-tastypie==0.14.3
djangorestframework==3.11.0
docutils==0.15.2
feedparser==5.1.3
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
future==0.18.2
guacamole==0.9.2
gunicorn==20.0.4
html5lib==0.999
httplib2==0.9.1
idna==2.0
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
joblib==0.14.1
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
murmurhash==1.0.2
nltk==3.5
numpy==1.18.3
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
pexpect==4.0.1
Pillow==7.1.1
plac==1.1.3
plainbox==0.25
preshed==3.0.2
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.0.3
python-apt==1.1.0b1+ubuntu0.16.4.8
python-dateutil==2.8.1
python-debian==0.1.27
python-mimeparse==1.6.0
python-systemd==231
pytz==2019.3
pyxdg==0.25
regex==2020.4.4
reportlab==3.3.0
requests==2.9.1
s3transfer==0.3.3
sacremoses==0.0.41
scikit-learn==0.22.2.post1
scipy==1.4.1
sentencepiece==0.1.86
sessioninstaller==0.0.0
six==1.10.0
spacy==2.2.4
sqlparse==0.3.1
srsly==1.0.2
ssh-import-id==5.5
system-service==0.3
textblob==0.15.3
thinc==7.4.0
tokenizers==0.5.2
torch==1.5.0+cpu
torchvision==0.6.0+cpu
tqdm==4.45.0
transformers==2.8.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.25.9
usb-creator==0.3.0
virtualenv==15.0.1
wasabi==0.6.0
Werkzeug==1.0.1
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
zipp==3.1.0
